@@ -9,19 +9,31 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "axios": "^0.19.0",
-    "expo": "^33.0.0",
+    "@react-native-community/masked-view": "0.1.6",
+    "@react-navigation/bottom-tabs": "^5.2.6",
+    "@react-navigation/native": "^5.1.5",
+    "@react-navigation/stack": "^5.2.10",
+    "axios": "^0.19.2",
+    "expo": "^37.0.0",
     "moment": "^2.24.0",
-    "react": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-33.0.0.tar.gz",
-    "react-native-extended-stylesheet": "^0.10.0",
-    "react-native-gesture-handler": "^1.3.0",
-    "react-navigation": "^3.11.0",
-    "react-redux": "^5.1.1",
-    "redux": "^4.0.1",
-    "socket.io-client": "^2.2.0"
+    "react": "16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.0.tar.gz",
+    "react-native-extended-stylesheet": "^0.12.0",
+    "react-native-gesture-handler": "~1.6.0",
+    "react-native-keyboard-aware-scroll-view": "^0.9.1",
+    "react-native-reanimated": "~1.7.0",
+    "react-native-safe-area-context": "0.7.3",
+    "react-native-screens": "~2.2.0",
+    "reactn": "^2.2.7",
+    "socket.io-client": "^2.3.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^5.0.0"
+    "babel-preset-expo": "^8.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^3.0.0",
+    "prettier": "^2.0.4"
   }
 }
