@@ -1,6 +1,9 @@
+// Libraries
 import React from 'react';
 import EStyleSheet from 'react-native-extended-stylesheet';
 import { AntDesign, Entypo, EvilIcons, Feather, FontAwesome, MaterialCommunityIcons } from '@expo/vector-icons';
+// String Constants
+import { STRING } from '../utils/stringConstants';
 
 const libraries = {
   AntDesign: AntDesign,
@@ -13,7 +16,7 @@ const libraries = {
 
 export default Icon = props => {
   const { color, library, name, size } = props;
-  const iconColor = typeof color === 'string' ? color : color();
+  const iconColor = typeof color === STRING ? color : color();
   const Library = libraries[library];
   return (
     <Library name={name} size={size} color={iconColor} />
