@@ -1,3 +1,4 @@
+// Libraries
 import React from 'react';
 import {
   Platform,
@@ -6,13 +7,15 @@ import {
   TouchableNativeFeedback,
   View
 } from 'react-native';
+// String Constants
+import { IOS } from '../utils/stringConstants';
 
 export default Touchable = props => {
   const iosTouchables = {
     highlight: TouchableHighlight,
     opacity: TouchableOpacity,
   };
-  const OSSpecificTouchable = Platform.OS === 'ios'
+  const OSSpecificTouchable = Platform.OS === IOS
                             ? iosTouchables[props.iosType]
                             : TouchableNativeFeedback;
   // console.log( Platform.OS );
