@@ -1,16 +1,20 @@
-import React from 'react';
+// Libraries
+import React from 'reactn';
 import { Image, Text, View } from 'react-native';
-import { connect } from 'react-redux';
 import EStyleSheet from 'react-native-extended-stylesheet';
-
+// Components
 import AnnouncementStripText from './AnnouncementStripText';
 import ImgSection from './ImgSection';
 import LikeButton from '../../components/LikeButton';
-
+// Helper Funcs
 import useAxios from '../../utils/axios-helpers';
 import { urlHostName } from '../../utils/global-variables';
+// String Constants
+import {
+  _SLASH, ANNOUNCEMENTS, HEART, HEART_OUTLINE, LIKE, MATERIAL_COMMUNITY_ICONS
+} from '../../utils/stringConstants';
 
-const path = `${urlHostName}/announcements/like`;
+const path = `${urlHostName}${_SLASH}${ANNOUNCEMENTS}${_SLASH}${LIKE}`;
 const { putWithAxios } = useAxios(path);
 
 class AnnouncementStrip extends React.Component {
@@ -44,7 +48,7 @@ class AnnouncementStrip extends React.Component {
     if (delta < DOUBLE_PRESS_DELAY) {
       this.updateLike({
         announcementId: this.props.announcement._id,
-        userId: this.props.user._id,
+        userId: this.global.user._id,
       });
     }
     this.lastPress = time;
@@ -67,7 +71,7 @@ class AnnouncementStrip extends React.Component {
       updateAnnouncement,
       updated
     } = this.props;
-    const userId = this.props.user && this.props.user._id;
+    const userId = this.global.user && this.global.user._id;
     const { _id, imgUrl, likes } = announcement;
     const liked = likes.includes(userId);
     return (
@@ -91,10 +95,10 @@ class AnnouncementStrip extends React.Component {
 
         <LikeButton
           addedStyle={{ width: imgWidth }}
-          library={{ liked: 'MaterialCommunityIcons', unliked: 'MaterialCommunityIcons' }}
+          library={{ liked: MATERIAL_COMMUNITY_ICONS, unliked: MATERIAL_COMMUNITY_ICONS }}
           liked={liked}
           likes={likes.length}
-          name={{ liked: 'heart', unliked: 'heart-outline' }}
+          name={{ liked: HEART, unliked: HEART_OUTLINE }}
           updateLike={() => this.updateLike({ announcementId: _id, userId })}
         />
 
@@ -132,10 +136,4 @@ const styles = EStyleSheet.create({
   }
 });
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-  };
-};
-
-export default connect(mapStateToProps)(AnnouncementStrip);
+export default AnnouncementStrip;
