@@ -1,15 +1,19 @@
-import React from 'react';
+// Libraries
+import React from 'reactn';
 import { Image, Text, View } from 'react-native';
-import { connect } from 'react-redux';
 import EStyleSheet from 'react-native-extended-stylesheet';
-
+// Components
 import LikeButton from '../../components/LikeButton';
 import Touchable from '../../components/Touchable';
-
+// Helper Funcs
 import useAxios from '../../utils/axios-helpers';
 import { urlHostName } from '../../utils/global-variables';
+// String Constants
+import {
+  _EMPTYSTRING, _SLASH, ANNOUNCEMENTS, LIKE, OPACITY,
+} from '../../utils/stringConstants';
 
-const path = `${urlHostName}/announcements/like`;
+const path = `${urlHostName}${_SLASH}${ANNOUNCEMENTS}${_SLASH}${LIKE}`;
 const { putWithAxios } = useAxios(path);
 
 class ImgSection extends React.Component {
@@ -43,7 +47,7 @@ class ImgSection extends React.Component {
   	if (delta < DOUBLE_PRESS_DELAY) {
       this.updateLike({
         announcementId: this.props.announcement._id,
-        userId: this.props.user._id,
+        userId: this.global.user._id,
       });
   	}
   	this.lastPress = time;
@@ -57,14 +61,14 @@ class ImgSection extends React.Component {
 
   render() {
     const { announcement, imgHeight, imgWidth } = this.props;
-    const userId = this.props.user ? this.props.user._id : '';
+    const userId = this.global.user ? this.global.user._id : _EMPTYSTRING;
     const { _id, imgUrl, likes } = announcement;
     const liked = likes.includes(userId);
     return (
       <View style={styles.imgWrap}>
         <Touchable
           activeOpacity={.5}
-          iosType='opacity'
+          iosType={OPACITY}
           onPress={this.handleDoublePress}
           style={{height: imgHeight, width: imgWidth}}
         >
@@ -94,10 +98,4 @@ const styles = EStyleSheet.create({
   },
 });
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-  };
-};
-
-export default connect(mapStateToProps)(ImgSection);
+export default ImgSection;
