@@ -1,114 +1,115 @@
-import React from 'react';
-import { AsyncStorage, Button, ImageBackground, Text, TextInput, View } from 'react-native';
-import { connect } from 'react-redux';
+// Libraries
+import React, { useGlobal, useState } from 'reactn';
+import { Button, ImageBackground, Text, TextInput, View } from 'react-native';
+import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
 import EStyleSheet from 'react-native-extended-stylesheet';
-
+// Components
 import Touchable from '../../components/Touchable';
-
-import { liftUser } from '../../redux/modules/user';
-import { urlHostName, getColor, tokenName } from '../../utils/global-variables';
+// Helpers
+import { urlHostName, getColor } from '../../utils/global-variables';
 import useAxios from '../../utils/axios-helpers';
+import { getToken, setTokenOnDevice } from '../../utils/token-helpers';
+import { blackBG } from '../../style-sheet';
+import { _EMPTYSTRING, HEIGHT_$, WIDTH_$, YELLOW_$ } from '../../utils/stringConstants';
 
 const path = `${urlHostName}/user/login`;
 const { postWithAxios } = useAxios(path);
 
-class LoginScreen extends React.Component {
-  static navigationOptions = {
-    header: null,
-  };
-  constructor(props) {
-    super(props)
-    this.state = {
-      email: '',
-      password: '',
-    }
+const LoginScreen = props => {
+  // Global State
+  const [user, setUser] = useGlobal('user');
+  // State
+  const [email, setEmail] = useState(_EMPTYSTRING);
+  const [password, setPassword] = useState(_EMPTYSTRING);
+
+  // FAKE REMOVE
+  React.useEffect(() => {
+    postWithAxios({ email: 'q@q.com', password: 'password' }).then(result => {
+      result.data.user
+      ? handleSuccess({ user: result.data.user, token: result.data.token })
+      : handleErr(result.data._message);
+    });
+  }, []);
+
+  const handleSuccess = async ({ user, token }) => {
+    await setTokenOnDevice(token);
+    // props.liftUser({ user, token });
+    setUser({ self: user, token });
+    // props.navigation.navigate('Main');
   }
 
-  setToken = async token => {
-    try {
-      await AsyncStorage.setItem(tokenName, token);
-    } catch (err) {
-      console.log('err: ', err);
-    }
-  }
-
-  handleSuccess = async ({ user, token }) => {
-    await this.setToken(token);
-    this.props.liftUser({ user, token });
-    this.props.navigation.navigate('Main');
-  }
-
-  handleErr = errMsg => {
+  const handleErr = errMsg => {
     console.log('signup failed with err: ', errMsg);
   }
 
-  handleSubmit = () => {
-    const { email, password } = this.state;
+  const handleSubmit = () => {
     postWithAxios({ email, password }).then(result => {
       result.data.user
-        ? this.handleSuccess({ user: result.data.user, token: result.data.token })
-        : this.handleErr(result.data._message);
+        ? handleSuccess({ user: result.data.user, token: result.data.token })
+        : handleErr(result.data._message);
     });
   }
 
-  render() {
-    const { email, password } = this.state;
-    const yellow = () => EStyleSheet.value('$yellow');
-    // temp
-    const height = () => EStyleSheet.value('$height');
-    const width = () => EStyleSheet.value('$width');
-    const imgUri = `https://www.placecage.com/c/${width()}/${height()}`;
-    // end temp
-    return (
-      <View style={styles.page}>
-        <ImageBackground blurRadius={4} source={{uri: imgUri}} style={styles.imgBgStyle}>
+  const yellow = () => EStyleSheet.value(YELLOW_$);
+  // temp
+  const height = () => EStyleSheet.value(HEIGHT_$);
+  const width = () => EStyleSheet.value(WIDTH_$);
+  const imgUri = `https://www.placecage.com/c/${width()}/${height()}`;
+  // end temp
+  return (
+    <KeyboardAwareScrollView
+      behavior='padding'
+      contentContainerStyle={styles.page}
+      resetScrollToCoords={{ x: 0, y: 0 }}
+      style={{ backgroundColor: blackBG }}
+    >
+      <ImageBackground blurRadius={4} source={{uri: imgUri}} style={styles.imgBgStyle}>
 
-          <View style={styles.contentWrapper}>
+        <View style={styles.contentWrapper}>
 
-            <View style={styles.logoPlaceholder}></View>
+          <View style={styles.logoPlaceholder}></View>
 
-            <TextInput
-              autoCapitalize='none'
-              onChangeText={text => this.setState({ email: text })}
-              placeholder='Email'
-              placeholderTextColor={yellow()}
-              style={[ styles.text, styles.textInput, styles.inputMargin ]}
-              textContentType='emailAddress'
-              value={email}
-            />
+          <TextInput
+            autoCapitalize='none'
+            onChangeText={setEmail}
+            placeholder='Email'
+            placeholderTextColor={yellow()}
+            style={[ styles.text, styles.textInput, styles.inputMargin ]}
+            textContentType='emailAddress'
+            value={email}
+          />
 
-            <TextInput
-              autoCapitalize='none'
-              onChangeText={text => this.setState({ password: text })}
-              placeholder='Password'
-              placeholderTextColor={yellow()}
-              style={[ styles.text, styles.textInput, styles.inputMargin ]}
-              textContentType='password'
-              value={password}
-            />
+          <TextInput
+            autoCapitalize='none'
+            onChangeText={setPassword}
+            placeholder='Password'
+            placeholderTextColor={yellow()}
+            style={[ styles.text, styles.textInput, styles.inputMargin ]}
+            textContentType='password'
+            value={password}
+          />
 
-            <View style={[ styles.signupTextWrap, styles.inputMargin ]}>
-              <Touchable iosType='opacity' onPress={() => this.props.navigation.navigate('Signup')}>
-                <Text style={[ styles.text, styles.smallText, styles.signupTextLink ]}>Sign Up</Text>
-              </Touchable>
-            </View>
-
+          <View style={[ styles.signupTextWrap, styles.inputMargin ]}>
+            <Touchable iosType='opacity' onPress={() => props.navigation.navigate('Signup')}>
+              <Text style={[ styles.text, styles.smallText, styles.signupTextLink ]}>Sign Up</Text>
+            </Touchable>
           </View>
 
-          <Touchable
-            activeOpacity={.5}
-            iosType='highlight'
-            onPress={this.handleSubmit}
-            underlayColor={ getColor('yellow') }
-            style={styles.submitButton}
-          >
-            <Text style={[ styles.text, styles.submitButtonText ]}>Log In</Text>
-          </Touchable>
+        </View>
 
-        </ImageBackground>
-      </View>
-    );
-  }
+        <Touchable
+          activeOpacity={.5}
+          iosType='highlight'
+          onPress={handleSubmit}
+          underlayColor={ getColor('yellow') }
+          style={styles.submitButton}
+        >
+          <Text style={[ styles.text, styles.submitButtonText ]}>Log In</Text>
+        </Touchable>
+
+      </ImageBackground>
+    </KeyboardAwareScrollView>
+  );
 }
 
 const styles = EStyleSheet.create({
@@ -174,17 +175,8 @@ const styles = EStyleSheet.create({
   },
 });
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-    token: state.user.token,
-  };
+LoginScreen.navigationOptions = {
+  header: null,
 };
 
-const mapDispatchToProps = dispatch => {
-  return {
-    liftUser: ({ user, token }) => dispatch( liftUser({ user, token }) ),
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(LoginScreen);
+export default LoginScreen;
