@@ -1,16 +1,21 @@
-import React from 'react';
+// Libraries
+import React from 'reactn';
 import { Text, View } from 'react-native';
-import { connect } from 'react-redux';
 import EStyleSheet from 'react-native-extended-stylesheet';
 import moment from 'moment';
-
+// Components
 import LikeButton from '../../components/LikeButton';
 import Touchable from '../../components/Touchable';
-
+// Helper Funcs
 import useAxios from '../../utils/axios-helpers';
 import { urlHostName } from '../../utils/global-variables';
+// String Constants
+import {
+  _EMPTYSTRING, _SLASH,
+  EVENTS, HEART, HEART_OUTLINE, LIKE, MATERIAL_COMMUNITY_ICONS, MMM, OPACITY
+} from '../../utils/stringConstants';
 
-const path = `${urlHostName}/events/like`;
+const path = `${urlHostName}${_SLASH}${EVENTS}${_SLASH}${LIKE}`;
 const { putWithAxios } = useAxios(path);
 
 class DateBox extends React.Component {
@@ -49,7 +54,7 @@ class DateBox extends React.Component {
     if (delta < DOUBLE_PRESS_DELAY) {
       this.updateLike({
         eventId: this.props.event._id,
-        userId: this.props.user._id,
+        userId: this.global.user._id,
       });
     }
     this.lastPress = time;
@@ -62,17 +67,17 @@ class DateBox extends React.Component {
   }
 
   render() {
-    const { event, user } = this.props;
+    const { event } = this.props;
     const { likes } = event;
-    const userId = user ? user._id : '';
+    const userId = this.global.user ? this.global.user._id : EventLegend;
     const liked = likes.includes(userId);
     const startDate = moment(event.startDate);
     const date = startDate.date();
-    const month = moment(startDate).format('MMM');
+    const month = moment(startDate).format(MMM);
     return (
       <View style={ [styles.dateTile, { backgroundColor: this.props.color }] }>
         <Touchable
-          iosType='opacity'
+          iosType={OPACITY}
           onPress={this.handleDoublePress}
           viewStyle={styles.viewStyle}
         >
@@ -80,10 +85,10 @@ class DateBox extends React.Component {
           <Text style={styles.dateText}>{date}</Text>
         </Touchable>
         <LikeButton
-          library={{ liked: 'MaterialCommunityIcons', unliked: 'MaterialCommunityIcons' }}
+          library={{ liked: MATERIAL_COMMUNITY_ICONS, unliked: MATERIAL_COMMUNITY_ICONS }}
           liked={liked}
           likes={likes.length}
-          name={{ liked: 'heart', unliked: 'heart-outline' }}
+          name={{ liked: HEART, unliked: HEART_OUTLINE }}
           updateLike={() => this.updateLike({ eventId: event._id, userId })}
         />
       </View>
@@ -113,10 +118,4 @@ const styles = EStyleSheet.create({
   },
 });
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-  };
-};
-
-export default connect(mapStateToProps)(DateBox);
+export default DateBox;
