@@ -1,22 +1,35 @@
-import React from 'react';
+// Libraries
+import React from 'reactn';
 import { Button, ScrollView, RefreshControl, SectionList, Text, View } from 'react-native';
+// import { connect } from 'react-redux';
+import moment from 'moment'
+import io from 'socket.io-client'
 import EStyleSheet from 'react-native-extended-stylesheet';
-
-import EventsKey from './EventsKey';
+// Components
+import EventLegend from './EventLegend';
 import EventStrip from './EventStrip';
-
+// Keys
 import eventKeys from './event-keys';
-
+// Helper Funcs
 import { getIndex } from '../../utils/helpers';
 import useAxios from '../../utils/axios-helpers';
 import { urlHostName } from '../../utils/global-variables';
+// String Constants
+import {
+  _EQUALS, _SLASH, _QUESTION_MARK,
+  BYMONTH, EVENTS, EVENT_LIKE_UPDATE, MMMM, SORT, WIDTH_$, YELLOW_$, YYYY,
+} from '../../utils/stringConstants';
 
-const path = `${urlHostName}/events/bymonth`;
-const { getWithAxios } = useAxios(path);
+const path = `${_SLASH}${EVENTS}`
+const url = `${urlHostName}${path}`
+const urlSortByMonth = `${url}${_QUESTION_MARK}${SORT}${_EQUALS}${BYMONTH}`
+const { getWithAxios } = useAxios(urlSortByMonth);
+
+const eventSocket = io(url)
 
 const { getEventKeys, getEventTypes } = eventKeys();
 
-export default class EventsSubScreen extends React.Component {
+class EventsSubScreen extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -75,15 +88,41 @@ export default class EventsSubScreen extends React.Component {
   }
 
   componentDidMount() {
+    eventSocket.on(EVENT_LIKE_UPDATE, data => {
+      const { event, userId } = data
+      const eventMonth = moment(event.startDate).format(YYYY) === moment().year()
+                        ? momentedEvent.format(MMMM)
+                        : `${moment(event.startDate).format(MMMM)} ${moment(event.startDate).format(YYYY)}`
+      const userIdFromRedux = this.global.user._id
+      if (userId !== userIdFromRedux) {
+        const { events } = this.state
+        const updateMonthEvents = (month, updatedEvent) => {
+          const updatedEvents = month.events.map(mappedEvent => {
+            return mappedEvent._id === updatedEvent._id
+                                      ? updatedEvent
+                                      : mappedEvent
+          })
+          return { ...month, events: updatedEvents }
+        }
+        const newEvents = events.map(month => {
+          return month.month === eventMonth ? updateMonthEvents(month, event) : month
+        })
+        this.setState({ events: newEvents })
+      }
+    })
     getWithAxios().then(result => {
       this.setState({ events: result.data.sortedEvents, eventTypes: getEventTypes });
     });
   }
 
+  componentWillUnmount() {
+    eventSocket.close()
+  }
+
   render() {
     const { events, eventTypes, removedTypes } = this.state;
-    const width = () => EStyleSheet.value('$width');
-    const yellow = () => EStyleSheet.value('$yellow');
+    const width = () => EStyleSheet.value(WIDTH_$);
+    const yellow = () => EStyleSheet.value(YELLOW_$);
     const sections = events.map(this.createMonthSection);
     const sectionList = <SectionList
                           sections={sections}
@@ -100,8 +139,8 @@ export default class EventsSubScreen extends React.Component {
                           keyExtractor={item => item._id}
                         />;
     return (
-      <View style={[styles.screen, {width: width()}]}>
-        <EventsKey filterEventTypes={this.filterEventTypes} removedTypes={removedTypes} />
+      <View style={[styles.screen, { width: width() }]}>
+        <EventLegend filterEventTypes={this.filterEventTypes} removedTypes={removedTypes} />
         {sectionList}
       </View>
     );
@@ -130,3 +169,12 @@ const styles = EStyleSheet.create({
     fontSize: '30rem'
   },
 });
+
+// const mapStateToProps = state => {
+//   return {
+//     user: state.user.user,
+//   };
+// };
+
+// export default connect(mapStateToProps)(EventsSubScreen);
+export default EventsSubScreen;
