@@ -1,116 +1,114 @@
-import React from 'react';
+// Libraries
+import React, { useEffect, useGlobal, useState } from 'reactn';
 import { Dimensions, RefreshControl, ScrollView, StatusBar, Text, View } from 'react-native';
-import { connect } from 'react-redux';
 import io from 'socket.io-client'
 import EStyleSheet from 'react-native-extended-stylesheet';
-
-import AnnouncementStrip from './AnnouncementStrip';
-
+// Components
+import AnnouncementStrip from '../announcements/AnnouncementStrip';
+// Helper Funcs
 import { getIndex } from '../../utils/helpers';
 import useAxios from '../../utils/axios-helpers';
 import { urlHostName } from '../../utils/global-variables';
+// String Constants
+import {
+  _ID, _SLASH, ANNOUNCEMENTS, ANNOUCEMENT_LIKE_UPDATE, LIGHT_CONTENT, WIDTH_$, WINDOW, YELLOW_$
+} from '../../utils/stringConstants';
 
-const path = '/announcements'
+const path = `${_SLASH}${ANNOUNCEMENTS}`;
 const url = `${urlHostName}${path}`;
 const { getWithAxios } = useAxios(url);
 
-const screenWidth = Dimensions.get('window').width;
+const announcementsSocket = io(url)
 
-class HomeScreen extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      announcements: null,
-      refreshing: false,
-      updated: false,
-    };
-  }
+const screenWidth = Dimensions.get(WINDOW).width;
 
-  static navigationOptions = {
-    header: null,
-  };
-
-  onRefresh = () => {
-    this.setState({ refreshing: true });
+const HomeScreen = props => {
+  // Global State
+  const [user] = useGlobal('user');
+  // State
+  const [announcements, setAnnouncements] = useState(null);
+  const [refreshing, setRefreshing] = useState(false);
+  const [updated, setUpdated] = useState(false);
+  // Effects
+  useEffect(() => {
+    announcementsSocket.on(ANNOUCEMENT_LIKE_UPDATE, data => {
+      const { announcement, userId } = data
+      const userIdFromRedux = user._id
+      if (userId !== userIdFromRedux) {
+        const newAnnouncements = announcements.map(mappedAnnouncement => {
+          return mappedAnnouncement._id === announcement._id
+                                            ? announcement
+                                            : mappedAnnouncement
+        })
+        setAnnouncements(newAnnouncements);
+      }
+    });
     getWithAxios().then(result => {
-      this.setState({ announcements: result.data.announcements.reverse(), refreshing: false });
+      setAnnouncements(result.data.announcements.reverse());
+    });
+    return () => announcementsSocket.close();
+  });
+  // Functions
+  const onRefresh = () => {
+    setRefreshing(true);
+    getWithAxios().then(result => {
+      setAnnouncements(result.data.announcements.reverse());
+      setRefreshing(false);
     });
   }
-  updateAnnouncement = ({ announcementId, userId }) => {
-    const announcements = this.state.announcements.slice(0);
-    const idx = getIndex('_id', announcements, announcementId);
+  const updateAnnouncement = ({ announcementId, userId }) => {
+    const updatedAnnouncements = announcements.slice(0);
+    const idx = getIndex(_ID, announcements, announcementId);
     const announcement = announcements[idx];
     const { likes } = announcement;
     likes.includes(userId)
       ? likes.splice( likes.indexOf(userId), 1 )
       : likes.push(userId);
-    this.setState({ announcements, updated: true });
+    setAnnouncements(updatedAnnouncements);
+    setUpdated(true);
   }
+  // Variables
+  const width = () => EStyleSheet.value(WIDTH_$);
+  const yellow = () => EStyleSheet.value(YELLOW_$);
+  const imgWidth = width() * .8;
+  const imgHeight = imgWidth / 15 * 8;
+  const padding = width() * .1;
+  // Components
+  const announcementStrips = announcements &&
+                        announcements.map((announcement, i) => {
+                          return (
+                            <AnnouncementStrip
+                              announcement={announcement}
+                              finishUpdate={() => setUpdated(false)}
+                              imgHeight={imgHeight}
+                              imgWidth={imgWidth}
+                              key={announcement._id}
+                              padding={padding}
+                              updateAnnouncement={updateAnnouncement}
+                              updated={updated}
+                            />
+                          );
+                        });
+  return (
+    <View style={styles.screen}>
+      <StatusBar barStyle={LIGHT_CONTENT} />
+      <Text style={[ styles.headerText, { paddingLeft: padding } ]}>What's New</Text>
 
-  componentDidMount() {
-    const announcementsSocket = io(url)
-    announcementsSocket.on('announcementLikeUpdate', data => {
-      const { announcement, userId } = data
-      const userIdFromRedux = this.props.user._id
-      if (userId !== userIdFromRedux) {
-        const { announcements } = this.state
-        const newAnnouncements = announcements
-                                  .filter(currentAnnouncement => currentAnnouncement._id !== announcement._id)
-                                  .concat(announcement)
-        this.setState({ announcements: newAnnouncements })
-      }
-    })
-    getWithAxios().then(result => {
-      this.setState({ announcements: result.data.announcements.reverse() });
-    });
-  }
+      <ScrollView
+        refreshControl={
+          <RefreshControl
+            colors={[yellow]}
+            onRefresh={onRefresh}
+            refreshing={refreshing}
+            tintColor={yellow()}
+          />
+        }
+      >
+        {announcementStrips}
+      </ScrollView>
 
-  componentWillUnmount() {
-    announcementsSocket.close()
-  }
-
-  render() {
-    const width = () => EStyleSheet.value('$width');
-    const yellow = () => EStyleSheet.value('$yellow');
-    const imgWidth = width() * .8;
-    const imgHeight = imgWidth / 15 * 8;
-    const padding = width() * .1;
-    const announcements = this.state.announcements &&
-                          this.state.announcements.map((announcement, i) => {
-                            return (
-                              <AnnouncementStrip
-                                announcement={announcement}
-                                finishUpdate={() => this.setState({ updated: false})}
-                                imgHeight={imgHeight}
-                                imgWidth={imgWidth}
-                                key={announcement._id}
-                                padding={padding}
-                                updateAnnouncement={this.updateAnnouncement}
-                                updated={this.state.updated}
-                              />
-                            );
-                          });
-    return (
-      <View style={styles.screen}>
-        <StatusBar barStyle='light-content' />
-        <Text style={[ styles.headerText, {paddingLeft: padding } ]}>What's New</Text>
-
-        <ScrollView
-          refreshControl={
-            <RefreshControl
-              colors={[yellow]}
-              onRefresh={this.onRefresh}
-              refreshing={this.state.refreshing}
-              tintColor={yellow()}
-            />
-          }
-        >
-          {announcements}
-        </ScrollView>
-
-      </View>
-    );
-  };
+    </View>
+  );
 };
 
 const styles = EStyleSheet.create({
@@ -131,10 +129,8 @@ const styles = EStyleSheet.create({
   },
 });
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-  };
+HomeScreen.navigationOptions = {
+  header: null,
 };
 
-export default connect(mapStateToProps)(HomeScreen);
+export default HomeScreen;
