@@ -1,9 +1,12 @@
 // Libraries
-import EStyleSheet from 'react-native-extended-stylesheet'
+import EStyleSheet from 'react-native-extended-stylesheet';
 // Components
-import WodSubScreen from './WodSubScreen';
+import EventsScreen from '../events/EventsScreen';
+import HomeScreen from './HomeScreen';
 // Style Sheet
 import { width } from '../../style-sheet';
+// String Constants
+import { _$ANNOUNCEMENTS, _$EVENTS } from '../../utils/stringConstants';
 
 const getBounds = idx => {
   const multiplier = .5 + idx;
@@ -14,9 +17,15 @@ const getBounds = idx => {
   }
 }
 
-const wodPages = [
-  { title: 'This week' },
-  { title: 'Past WODs' },
+const homePages = [
+  {
+    component: HomeScreen,
+    title: _$ANNOUNCEMENTS,
+  },
+  {
+    component: EventsScreen,
+    title: _$EVENTS,
+  },
 ].map((page, i) => {
   const { lower, lowerVisible, upper } = getBounds(i);
   page.lowerScrollBounds = lower;
@@ -25,4 +34,4 @@ const wodPages = [
   return page;
 });
 
-export default wodPages;
+export default homePages;
